Use large integers for tracking data field

+++ tracking/models.py
@@ -135,7 +135,7 @@ class TrackingEvent(models.Model):
     name = models.CharField(max_length=50, help_text=_('Name identifying the event'))
     visitor = models.ForeignKey(Visitor, help_text=_('The visitor which generated the event'))
     time = models.DateTimeField(auto_now_add=True)
-    data = models.IntegerField(null=True, help_text=_('A name-specific identifier providing more information on the event'))
+    data = models.BigIntegerField(null=True, help_text=_('A name-specific identifier providing more information on the event'))
 
     EVENT_CHOICES = (
         ('click', 'Click'),
